moved follow notification filter to deer settings
