+++ src/screens/Settings/DeerSettings.tsx
@@ -24,6 +24,10 @@ import {
   useSetDirectFetchRecords,
 } from '#/state/preferences/direct-fetch-records'
 import {
+  useHideFollowNotifications,
+  useSetHideFollowNotifications,
+} from '#/state/preferences/hide-follow-notifications'
+import {
   useNoAppLabelers,
   useSetNoAppLabelers,
 } from '#/state/preferences/no-app-labelers'
@@ -43,6 +47,7 @@ import {Button, ButtonText} from '#/components/Button'
 import * as Dialog from '#/components/Dialog'
 import * as Toggle from '#/components/forms/Toggle'
 import {Atom_Stroke2_Corner0_Rounded as DeerIcon} from '#/components/icons/Atom'
+import {Bell_Stroke2_Corner0_Rounded as BellIcon} from '#/components/icons/Bell'
 import {Eye_Stroke2_Corner0_Rounded as VisibilityIcon} from '#/components/icons/Eye'
 import {Earth_Stroke2_Corner2_Rounded as GlobeIcon} from '#/components/icons/Globe'
 import {Lab_Stroke2_Corner0_Rounded as BeakerIcon} from '#/components/icons/Lab'
@@ -135,6 +140,9 @@ export function DeerSettingsScreen({}: Props) {
 
   const noDiscoverFallback = useNoDiscoverFallback()
   const setNoDiscoverFallback = useSetNoDiscoverFallback()
+
+  const hideFollowNotifications = useHideFollowNotifications()
+  const setHideFollowNotifications = useSetHideFollowNotifications()
 
   const location = useGeolocation()
   const setLocationControl = Dialog.useDialogControl()
@@ -309,6 +317,24 @@ export function DeerSettingsScreen({}: Props) {
               style={[a.w_full]}>
               <Toggle.LabelText style={[a.flex_1]}>
                 <Trans>Do not fall back to discover feed</Trans>
+              </Toggle.LabelText>
+              <Toggle.Platform />
+            </Toggle.Item>
+          </SettingsList.Group>
+
+          <SettingsList.Group contentContainerStyle={[a.gap_sm]}>
+            <SettingsList.ItemIcon icon={BellIcon} />
+            <SettingsList.ItemText>
+              <Trans>Notification Filters</Trans>
+            </SettingsList.ItemText>
+            <Toggle.Item
+              name="hide_follow_notifications"
+              label={_(msg`Hide follow notifications`)}
+              value={hideFollowNotifications ?? false}
+              onChange={value => setHideFollowNotifications(value)}
+              style={[a.w_full]}>
+              <Toggle.LabelText style={[a.flex_1]}>
+                <Trans>Hide follow notifications</Trans>
               </Toggle.LabelText>
               <Toggle.Platform />
             </Toggle.Item>
